src/cortex/requirements.txt
# Chyren dependencies
# Core uses Python stdlib only (json, hmac, hashlib, urllib, threading)
# python-dotenv is optional — Chyren loads one-true.env directly in main.py
# Add provider SDKs here only if switching from urllib to official client libraries
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Multi-Modal Ingestion Pipeline
datasets>=2.18.0
transformers>=4.38.0
huggingface-hub>=0.21.0
evaluate>=0.4.1
bert-score>=0.3.13
accelerate>=0.27.0
peft>=0.9.0
qdrant-client>=1.8.0
psycopg2-binary>=2.9.9
mcp>=1.0.0
websockets>=12.0.0
langgraph>=0.0.10
langchain-core>=0.1.0
